Consolidated to coord.py, orbel.py, geometry.py.
