+++ phoebe/dynamics/geometry.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 
 import numpy as np
-from phoebe.dynamics import coord_h2j
-from phoebe.dynamics import coord_j2b
-from phoebe.dynamics import orbel_el2xv
+
+from phoebe.dynamics import coord
+from phoebe.dynamics import orbel
 
 def geometry(m, elmts, geometry='hierarchical'):
     """
@@ -13,11 +13,11 @@ def geometry(m, elmts, geometry='hierarchical'):
 
     if geometry == 'hierarchical':
 
-        _geometry = hierarchical
+        _geometry = geometry_hierarchical
 
     elif geometry == 'twopairs':
 
-        _geometry = twopairs
+        _geometry = geometry_twopairs
 
     else:
         raise NotImplementedError
@@ -47,7 +47,7 @@ def geometry(m, elmts, geometry='hierarchical'):
 
     return xs, ys, zs, vxs, vys, vzs
 
-def hierarchical(m, elmts):
+def geometry_hierarchical(m, elmts):
     """
      _          \                |
     / \          |               |
@@ -66,14 +66,14 @@ def hierarchical(m, elmts):
         msum += m[j]
         ialpha = -1
 
-        rj[j], vj[j] = orbel_el2xv.orbel_el2xv(msum, ialpha, elmts[j-1])
+        rj[j], vj[j] = orbel.orbel_el2xv(msum, ialpha, elmts[j-1])
 
     # convert to barycentric frame
-    rb, vb = coord_j2b.coord_j2b(m, rj, vj)
+    rb, vb = coord.coord_j2b(m, rj, vj)
 
     return rb, vb
 
-def twopairs(m, elmts):
+def geometry_twopairs(m, elmts):
     """
      _          _               \ 
     / \        / \               |
@@ -91,7 +91,7 @@ def twopairs(m, elmts):
     # (1+2) pair, 1-centric coordinates
     msum = m[0]+m[1]
     ialpha = -1
-    r2_1, v2_1 = orbel_el2xv.orbel_el2xv(msum, ialpha, elmts[0])
+    r2_1, v2_1 = orbel.orbel_el2xv(msum, ialpha, elmts[0])
 
     # barycenter
     r12_1 = m[1]/msum * r2_1
@@ -99,7 +99,7 @@ def twopairs(m, elmts):
 
     # (3+4) pair, 3-centric
     msum = m[2]+m[3]
-    r4_3, v4_3 = orbel_el2xv.orbel_el2xv(msum, ialpha, elmts[1])
+    r4_3, v4_3 = orbel.orbel_el2xv(msum, ialpha, elmts[1])
 
     # barycenter
     r34_3 = m[3]/msum * r4_3
@@ -107,7 +107,7 @@ def twopairs(m, elmts):
 
     # (1+2)+(3+4) mutual orbit, (1+2)-centric
     msum = m[0]+m[1]+m[2]+m[3]
-    r34_12, v34_12 = orbel_el2xv.orbel_el2xv(msum, ialpha, elmts[2])
+    r34_12, v34_12 = orbel.orbel_el2xv(msum, ialpha, elmts[2])
 
     # everything to 1-centric
     rh[0,:] = 0.0
@@ -120,55 +120,152 @@ def twopairs(m, elmts):
     vh[3,:] = vh[2,:] + v4_3
 
     # everything to Jacobian
-    rj[0:4], vj[0:4] = coord_h2j.coord_h2j(m[0:4], rh[0:4], vh[0:4])
+    rj[0:4], vj[0:4] = coord.coord_h2j(m[0:4], rh[0:4], vh[0:4])
 
     # other bodies (also Jacobian)
     for j in range(4, nbod):
         msum += m[j]
 
-        rj[j], vj[j] = orbel_el2xv.orbel_el2xv(msum, ialpha, elmts[j-1])
+        rj[j], vj[j] = orbel.orbel_el2xv(msum, ialpha, elmts[j-1])
 
     # convert to barycentric frame
-    rb, vb = coord_j2b.coord_j2b(m, rj, vj)
+    rb, vb = coord.coord_j2b(m, rj, vj)
 
     return rb, vb
 
-if __name__ == "__main__":
+########################################################################
 
-    day = 86400.
-    au = 1.496e11
-    M_S = 2.e30
-    G = 6.67e-11
-    gms = G*M_S / (au**3 * day**-2)
+def invgeometry(m, rb, vb, geometry='hierarchical'):
+    """
+    Convert barycentric coordinates to elements for various geometries.
+   
+    """
 
-    m = gms*np.array([1.0, 0.0])
-    elmts = [[1.0, 0.0, 0.0, 0.0, 0.0, 0.0]]
+    if geometry == 'hierarchical':
 
-    rb, vb = hierarchical(m, elmts)
+        _invgeometry = invgeometry_hierarchical
 
-    print("m = ", m)
-    print("rb[0] = ", rb[0])
-    print("rb[1] = ", rb[1])
-    print("vb[0] = ", vb[0])
-    print("vb[1] = ", vb[1])
+    elif geometry == 'twopairs':
 
-    m = gms*np.array([1.0, 1.0, 0.5, 0.5])
-    elmts = []
-    elmts.append([0.1, 0.0, 0.0, 0.0, 0.0, 0.0])
-    elmts.append([0.2, 0.0, 0.0, 0.0, 0.0, 0.0])
-    elmts.append([1.0, 0.0, 0.0, 0.0, 0.0, 0.0])
+        _invgeometry = invgeometry_twopairs
 
-    rb, vb = twopairs(m, elmts)
+    else:
+        raise NotImplementedError
 
-    print("")
-    print("m = ", m)
-    print("rb[0] = ", rb[0])
-    print("rb[1] = ", rb[1])
-    print("rb[2] = ", rb[2])
-    print("rb[3] = ", rb[3])
-    print("vb[0] = ", vb[0])
-    print("vb[1] = ", vb[1])
-    print("vb[2] = ", vb[2])
-    print("vb[3] = ", vb[3])
+    # cf. mutable numpy arrays
+    rb = np.copy(rb)
+    vb = np.copy(vb)
+
+    # orientation
+    rb[:,0] *= -1.0
+    rb[:,1] *= -1.0
+    vb[:,0] *= -1.0
+    vb[:,1] *= -1.0
+
+    return _invgeometry(m, rb, vb)
+
+def invgeometry_hierarchical(m, rb, vb):
+    """
+     _          \                |
+    / \          |               |
+    1 2          3               4
+    \_/          |               |
+                /                |
+    """
+
+    nbod = len(m)
+    elmts = np.zeros((nbod-1, 6))
+    euler = np.zeros((nbod, 3))
+    roche = np.zeros((nbod, 2))
+
+    # convert to Jacobian coordinates
+    rj, vj = coord.coord_b2j(m, rb, vb)
+
+    # compute osculating elements
+    msum = m[0]
+    for j in range(1, nbod):
+        msum += m[j]
+        ialpha = -1
+
+        elmts[j-1,:] = orbel.orbel_xv2el(msum, rj[j], vj[j])
+        euler[j,:] = orbel.get_euler(elmts[j-1])
+        roche[j,:] = orbel.get_roche(msum, elmts[j-1])
+
+        if j==1:
+            euler[0,:] = euler[1,:]
+            roche[0,:] = roche[1,:]
+        if j>=1:
+            euler[j,0] += np.pi
+
+    return elmts, euler, roche
+
+def invgeometry_twopairs(m, rb, vb):
+    """
+     _          _               \ 
+    / \        / \               |
+    1 2        3 4               5
+    \_/        \_/               |
+                                / 
+    """
+
+    nbod = len(m)
+    elmts = np.zeros((nbod-1, 6))
+    euler = np.zeros((nbod, 3))
+    roche = np.zeros((nbod, 2))
+
+    # (1+2) pair, 1-centric coordinates
+    msum = m[0]+m[1]
+    r2_1 = rb[1] - rb[0]
+    v2_1 = vb[1] - vb[0]
+
+    elmts[0,:] = orbel.orbel_xv2el(msum, r2_1, v2_1)
+    euler[0,:] = orbel.get_euler(elmts[0])
+    roche[0,:] = orbel.get_roche(msum, elmts[0])
+
+    euler[1,:] = euler[0,:]
+    roche[1,:] = roche[0,:]
+    euler[1,0] += np.pi
+
+    # barycenter
+    r12 = (m[0]*rb[0] + m[1]*rb[1])/msum
+    v12 = (m[0]*vb[0] + m[1]*vb[1])/msum
+
+    # (3+4) pair, 3-centric
+    msum = m[2]+m[3]
+    r4_3 = rb[3] - rb[2]
+    v4_3 = vb[3] - vb[2]
+
+    elmts[1,:] = orbel.orbel_xv2el(msum, r4_3, v4_3)
+    euler[2,:] = orbel.get_euler(elmts[1])
+    roche[2,:] = orbel.get_roche(msum, elmts[1])
+    
+    euler[3,:] = euler[2,:]
+    roche[3,:] = roche[2,:]
+    euler[3,0] += np.pi
+
+    # barycenter
+    r34 = (m[2]*rb[2] + m[3]*rb[3])/msum
+    v34 = (m[2]*vb[2] + m[3]*vb[3])/msum
+
+    # (1+2)+(3+4) mutual orbit, (1+2)-centric
+    msum = m[0]+m[1]+m[2]+m[3]
+    r34_12 = r34 - r12
+    v34_12 = v34 - v12
+    elmts[2,:] = orbel.orbel_xv2el(msum, r34_12, v34_12)
+
+    # everything to Jacobian
+    rj, vj = coord.coord_b2j(m, rb, vb)
+    
+    # other bodies (also Jacobian)
+    for j in range(4, nbod):
+        msum += m[j]
+
+        elmts[j-1,:] = orbel.orbel_xv2el(msum, rj[j], vj[j])
+        euler[j,:] = orbel.get_euler(elmts[j-1])
+        roche[j,:] = orbel.get_roche(msum, elmts[j-1])
+
+        euler[j,0] += np.pi
+
+    return elmts, euler, roche
